Rename files, order code
- rename files caption-button.js -> toggle-button.js and video-transcript.js -> videojs-transcript.js
- remove object Component
- rewrite method createEl
