+++ video_xblock/backends/base.py
@@ -69,10 +69,10 @@ class BaseVideoPlayer(Plugin):
             self.render_resource('../static/js/videojs-tabindex.js', **context)
         )
         frag.add_javascript(
-            self.resource_string('../static/js/caption-button.js')
+            self.resource_string('../static/js/toggle-button.js')
         )
         frag.add_javascript(
-            self.render_resource('../static/js/video-transcript.js', **context)
+            self.render_resource('../static/js/videojs-transcript.js', **context)
         )
         frag.add_javascript(self.render_resource(
             '../static/js/videojs_event_plugin.js', **context
