+++ video_xblock/backends/brightcove.py
@@ -38,9 +38,9 @@ class BrightcovePlayer(BaseVideoPlayer):
         frag.add_javascript(
             self.render_resource('../static/js/videojs-tabindex.js', **context)
         )
-        frag.add_javascript(
-            self.resource_string('../static/js/caption-button.js')
-        )
+        # frag.add_javascript(
+        #     self.resource_string('../static/js/toggle-button.js')
+        # )
         frag.add_javascript(
             self.render_resource('../static/js/video-transcript.js', **context)
         )
